Moving PSA components to component folder
