+++ src/components/Auth/AuthButton/AuthButton.js
@@ -6,7 +6,7 @@
 import { Typography } from '@mui/material';
 import { useAuth0 } from '@auth0/auth0-react';
 import React from 'react';
-import PSAButton from '../../../shared/PSAButton';
+import PSAButton from '../../PSAComponents/PSAButton';
 
 const AuthButton = ({
   type, variant = 'text', onClickCallback = () => {}, color = 'secondary',
