+++ src/pages/Landing/AuthModal/AuthModal.js
@@ -8,7 +8,7 @@ import {
 import React, { useEffect } from 'react';
 import { useAuth0 } from '@auth0/auth0-react';
 import AuthButton from '../../../components/Auth/AuthButton/AuthButton';
-import PSAModal from '../../../shared/PSAModal';
+import PSAModal from '../../../components/PSAComponents/PSAModal';
 
 const localStorageKey = 'notShowAuth';
