+++ src/shared/ProgressButtonsInner.js
@@ -10,7 +10,7 @@ import { useSelector, useDispatch } from 'react-redux';
 import { useHistory } from 'react-router-dom';
 import { reset } from '../reduxStore/store';
 import { updateProgress, setMyCoverCropReset } from '../reduxStore/sharedSlice';
-import PSAButton from './PSAButton';
+import PSAButton from '../components/PSAComponents/PSAButton';
 
 const ProgressButtonsInner = ({
   isDisabledBack, isDisabledNext, isDisabledRefresh, toolTip,
